# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Nami Transaction Tracking System.",
        "title": "Nami Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/actions": {
            "post": {
                "description": "Perform a domain-specific action",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Perform an action",
                "parameters": [
                    {
                        "description": "Action request",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ActionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ActionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/transactions": {
            "get": {
                "description": "Get a list of transactions or create a new one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List or create transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated transaction types",
                        "name": "types",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated asset symbols",
                        "name": "assets",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated account names",
                        "name": "accounts",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tag names",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Counterparty name",
                        "name": "counterparty",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currencies for FX conversion (comma-separated), defaults to USD,VND",
                        "name": "currencies",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basic transactions without FX conversion",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Get a list of transactions or create a new one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List or create transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated transaction types",
                        "name": "types",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated asset symbols",
                        "name": "assets",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated account names",
                        "name": "accounts",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated tag names",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Counterparty name",
                        "name": "counterparty",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currencies for FX conversion (comma-separated), defaults to USD,VND",
                        "name": "currencies",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Basic transactions without FX conversion",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "description": "Operate on a single transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get, update, or delete a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Operate on a single transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get, update, or delete a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Operate on a single transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get, update, or delete a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/vaults/{vaultId}/transactions": {
            "get": {
                "description": "List transactions for a vault (supports limit, offset, type, status)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "List vault transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VaultTransaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new transaction recorded in the vault ledger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "Create a vault transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vault transaction payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VaultTransaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.VaultTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/vaults/{vaultId}/user/{userId}/transactions": {
            "get": {
                "description": "List transactions for a user in a vault (supports limit, offset)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaults"
                ],
                "summary": "List user transactions in a vault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vault ID",
                        "name": "vaultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VaultTransaction"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ActionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "models.ActionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "executed_at": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
...
        "models.Transaction": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "amount_local": {
                    "description": "Amount fields",
                    "type": "number"
                },
                "asset": {
                    "type": "string"
                },
                "borrow_active": {
                    "type": "boolean"
                },
                "borrow_apr": {
                    "description": "Borrow metadata (for type = 'borrow')",
                    "type": "number"
                },
                "borrow_term_days": {
                    "type": "integer"
                },
                "cashflow_local": {
                    "type": "number"
                },
                "counterparty": {
                    "type": "string"
                },
                "created_at": {
                    "description": "Audit fields",
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "delta_qty": {
                    "description": "Derived metrics",
                    "type": "number"
                },
                "entry_date": {
                    "type": "string"
                },
                "exit_date": {
                    "type": "string"
                },
                "fee_local": {
                    "description": "Fees (stored in local currency, will be converted dynamically)",
                    "type": "number"
                },
                "horizon": {
                    "description": "Optional tracking",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "internal_flow": {
                    "description": "Flow flags",
                    "type": "boolean"
                },
                "investment_id": {
                    "description": "Enhanced investment tracking - links transaction to investment position",
                    "type": "string"
                },
                "local_currency": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "price_local": {
                    "type": "number"
                },
                "quantity": {
                    "description": "Amount fields",
                    "type": "number"
                },
                "tag": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}